# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Active Impact Talent Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get ecosystem stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/candidates:
    post:
      operationId: createCandidate
      tags: [admin]
      summary: Create a new candidate (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCandidate"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/candidates/{id}:
    patch:
      operationId: updateCandidate
      tags: [admin]
      summary: Update a candidate (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCandidate"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCandidate
      tags: [admin]
      summary: Delete a candidate (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/domains:
    get:
      operationId: listDomains
      tags: [admin]
      summary: List whitelisted domains (admin only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Domain"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addDomain
      tags: [admin]
      summary: Add a whitelisted domain (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Domain already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/domains/{id}:
    delete:
      operationId: deleteDomain
      tags: [admin]
      summary: Remove a whitelisted domain (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/intro-requests:
    get:
      operationId: listAdminIntroRequests
      tags: [admin]
      summary: List all intro requests with full candidate and founder info (admin only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminIntroRequest"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AdminCandidate:
      type: object
      properties:
        id:
          type: integer
        internalId:
          type: string
        realName:
          type: string
        anonymizedHeadline:
          type: string
        roleCategory:
          type: string
          enum: [Engineering, Sales, Operations, Product, Finance, Marketing, Executive]
        seniority:
          type: string
          enum: [IC, Manager, Director, VP, C-level]
        yearsExperience:
          type: integer
        location:
          type: string
        openToRelocation:
          type: boolean
        compRangeMin:
          type: integer
        compRangeMax:
          type: integer
        topSkills:
          type: array
          items:
            type: string
        summaryBlurb:
          type: string
        notableCredentials:
          type: string
        status:
          type: string
          enum: [opted_in, paused, placed, withdrawn]
        dateAdded:
          type: string
          format: date
      required:
        - id
        - internalId
        - realName
        - anonymizedHeadline
        - roleCategory
        - seniority
        - yearsExperience
        - location
        - openToRelocation
        - compRangeMin
        - compRangeMax
        - topSkills
        - summaryBlurb
        - notableCredentials
        - status
        - dateAdded
    CandidateInput:
      type: object
      properties:
        internalId:
          type: string
        realName:
          type: string
        anonymizedHeadline:
          type: string
        roleCategory:
          type: string
          enum: [Engineering, Sales, Operations, Product, Finance, Marketing, Executive]
        seniority:
          type: string
          enum: [IC, Manager, Director, VP, C-level]
        yearsExperience:
          type: integer
        location:
          type: string
        openToRelocation:
          type: boolean
        compRangeMin:
          type: integer
        compRangeMax:
          type: integer
        topSkills:
          type: array
          items:
            type: string
        summaryBlurb:
          type: string
        notableCredentials:
          type: string
        status:
          type: string
          enum: [opted_in, paused, placed, withdrawn]
        dateAdded:
          type: string
          format: date
      required:
        - internalId
        - realName
        - anonymizedHeadline
        - roleCategory
        - seniority
        - yearsExperience
        - location
        - openToRelocation
        - compRangeMin
        - compRangeMax
        - topSkills
        - summaryBlurb
        - notableCredentials
        - status
    CandidateUpdate:
      type: object
      properties:
        realName:
          type: string
        anonymizedHeadline:
          type: string
        roleCategory:
          type: string
          enum: [Engineering, Sales, Operations, Product, Finance, Marketing, Executive]
        seniority:
          type: string
          enum: [IC, Manager, Director, VP, C-level]
        yearsExperience:
          type: integer
        location:
          type: string
        openToRelocation:
          type: boolean
        compRangeMin:
          type: integer
        compRangeMax:
          type: integer
        topSkills:
          type: array
          items:
            type: string
        summaryBlurb:
          type: string
        notableCredentials:
          type: string
        status:
          type: string
          enum: [opted_in, paused, placed, withdrawn]
...
    AdminIntroRequest:
      type: object
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        candidateHeadline:
          type: string
        candidateRealName:
          type: string
        candidateRoleCategory:
          type: string
        founderId:
          type: integer
        founderName:
          type: string
        founderEmail:
          type: string
        founderCompany:
          type: ["string", "null"]
        status:
          type: string
          enum: [requested, offered, intro_made, placed, closed]
        requestedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - candidateId
        - candidateHeadline
        - candidateRealName
        - candidateRoleCategory
        - founderId
        - founderName
        - founderEmail
        - status
        - requestedAt
        - updatedAt
...
    AdminStats:
      type: object
      properties:
        totalCandidates:
          type: integer
        candidatesOffered:
          type: integer
        introsMade:
          type: integer
        placements:
          type: integer
      required:
        - totalCandidates
        - candidatesOffered
        - introsMade
        - placements
    Domain:
      type: object
      properties:
        id:
          type: integer
        domain:
          type: string
        companyName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - domain
        - createdAt
    DomainInput:
      type: object
      properties:
        domain:
          type: string
        companyName:
          type: string
      required:
        - domain