# lib/api-zod/src/generated/types/adminCandidate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Active Impact Talent Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminCandidateRoleCategory } from './adminCandidateRoleCategory';
import type { AdminCandidateSeniority } from './adminCandidateSeniority';
import type { AdminCandidateStatus } from './adminCandidateStatus';

export interface AdminCandidate {
  id: number;
  internalId: string;
  realName: string;
  anonymizedHeadline: string;
  roleCategory: AdminCandidateRoleCategory;
  seniority: AdminCandidateSeniority;
  yearsExperience: number;
  location: string;
  openToRelocation: boolean;
  compRangeMin: number;
  compRangeMax: number;
  topSkills: string[];
  summaryBlurb: string;
  notableCredentials: string;
  status: AdminCandidateStatus;
  dateAdded: Date;
}