# lib/api-zod/src/generated/types/adminIntroRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Active Impact Talent Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminIntroRequestStatus } from './adminIntroRequestStatus';

export interface AdminIntroRequest {
  id: number;
  candidateId: number;
  candidateHeadline: string;
  candidateRealName: string;
  candidateRoleCategory: string;
  founderId: number;
  founderName: string;
  founderEmail: string;
  /** @nullable */
  founderCompany?: string | null;
  status: AdminIntroRequestStatus;
  requestedAt: Date;
  updatedAt: Date;
}