# lib/api-zod/src/generated/types/candidateUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Active Impact Talent Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateUpdateRoleCategory } from './candidateUpdateRoleCategory';
import type { CandidateUpdateSeniority } from './candidateUpdateSeniority';
import type { CandidateUpdateStatus } from './candidateUpdateStatus';

export interface CandidateUpdate {
  realName?: string;
  anonymizedHeadline?: string;
  roleCategory?: CandidateUpdateRoleCategory;
  seniority?: CandidateUpdateSeniority;
  yearsExperience?: number;
  location?: string;
  openToRelocation?: boolean;
  compRangeMin?: number;
  compRangeMax?: number;
  topSkills?: string[];
  summaryBlurb?: string;
  notableCredentials?: string;
  status?: CandidateUpdateStatus;
}