# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Active Impact Talent Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Domain {
  id: number;
  domain: string;
  /** @nullable */
  companyName?: string | null;
  createdAt: Date;
}